Developer/sojib4

+++ src/component/ui/AdminDashboard/OrderListCard.tsx
@@ -1,6 +1,11 @@
 import { MdOutlineDelete, MdOutlineSecurityUpdate } from "react-icons/md";
-import { useDeleteOrderMutation } from "../../../redux/features/api/endpoints/orderApi";
+import {
+  useDeleteOrderMutation,
+  useUpdateOrderMutation,
+} from "../../../redux/features/api/endpoints/orderApi";
 import toast from "react-hot-toast";
+import Swal from "sweetalert2";
+import { SubmitHandler, useForm } from "react-hook-form";
 
 export type TTransaction = {
   id: string;
@@ -18,7 +23,7 @@ export type TOrder = {
   productId: string;
   quantity: number;
   totalPrice: number;
-  status: "Paid" | "Pending" | "Cancelled";
+  status: "Paid" | "Pending" | "Cancelled" | "Completed" | "Shipped";
   transaction: TTransaction;
   createdAt: string;
   updatedAt: string;
@@ -28,15 +33,61 @@ export type TOrder = {
 const OrderListCard = ({ order }: { order: TOrder }) => {
   const { _id, email, productId, quantity, totalPrice, status } = order;
   const [deleteOrder] = useDeleteOrderMutation();
+  const [updateOrder] = useUpdateOrderMutation();
+
+  const { register, handleSubmit, reset } = useForm<Partial<TOrder>>({
+    defaultValues: {
+      email,
+      productId,
+      quantity,
+      totalPrice,
+      status,
+    },
+  });
+
+  const onSubmit: SubmitHandler<Partial<TOrder>> = async (data) => {
+    console.log(_id, data);
+    const parsedData = {
+      ...data,
+      quantity: Number(data.quantity),
+      totalPrice: Number(data.totalPrice),
+    };
+    const res = await updateOrder({ id: _id, payload: parsedData });
+    if (res?.data?.success) {
+      Swal.fire({
+        position: "center",
+        icon: "success",
+        title: `${productId} successfully updated.`,
+        showConfirmButton: false,
+        timer: 1500,
+      });
+      reset();
+    }
+  };
 
   const handleDelete = async (id: string) => {
-    const res = await deleteOrder(id).unwrap();
-    console.log(res);
-    if (res.success) {
-      toast.success(`${res?.data?.productId} has been deleted!`);
-    } else {
-      toast.error(`${productId} can't be deleted right now!`);
-    }
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const res = await deleteOrder(id).unwrap();
+        if (res.success) {
+          Swal.fire({
+            title: "Deleted!",
+            text: `${res?.data?.productId} has been deleted!`,
+            icon: "success",
+          });
+        } else {
+          toast.error(`${productId} can't be deleted right now!`);
+        }
+      }
+    });
   };
   return (
     <div className="card bg-neutral text-neutral-content w-full">
@@ -59,8 +110,61 @@ const OrderListCard = ({ order }: { order: TOrder }) => {
           <input type="checkbox" id="my_modal_7" className="modal-toggle" />
           <div className="modal" role="dialog">
             <div className="modal-box">
-              <h3 className="text-lg font-bold">Hello!</h3>
-              <p className="py-4">This modal works with a hidden checkbox!</p>
+              <div className="hero bg-base-200">
+                <div className="card bg-base-100 w-full shrink-0 shadow-2xl">
+                  <h3 className="text-lg md:text-xl font-semibold text-center">
+                    Update Order
+                  </h3>
+                  <form onSubmit={handleSubmit(onSubmit)} className="card-body">
+                    <fieldset className="fieldset">
+                      <label className="fieldset-label">Email</label>
+                      <input
+                        type="email"
+                        className="input"
+                        placeholder="Email"
+                        {...register("email")}
+                      />
+                      <label className="fieldset-label">ProductId</label>
+                      <input
+                        type="text"
+                        className="input"
+                        placeholder="ProductId"
+                        {...register("productId")}
+                      />
+                      <label className="fieldset-label">Quantity</label>
+                      <input
+                        type="number"
+                        className="input"
+                        placeholder="Quantity"
+                        {...register("quantity")}
+                      />
+                      <label className="fieldset-label">Total Price</label>
+                      <input
+                        type="number"
+                        className="input"
+                        placeholder="Total Price"
+                        {...register("totalPrice")}
+                      />
+                      <label className="fieldset-label">Status</label>
+                      <select
+                        defaultValue="Pick a color"
+                        className="select"
+                        {...register("status")}
+                      >
+                        <option disabled={true}>Status</option>
+                        <option value="paid">Paid</option>
+                        <option value="pending">Pending</option>
+                        <option value="completed">Completed</option>
+                        <option value="cancelled">Cancelled</option>
+                        <option value="shipped">Shipped</option>
+                      </select>
+                      <button className="btn btn-primary mt-4">
+                        Update Order
+                      </button>
+                    </fieldset>
+                  </form>
+                </div>
+              </div>
             </div>
             <label className="modal-backdrop" htmlFor="my_modal_7">
               Close
